Utiltize libcnb.Binding.SecretFilePath
Previously, there was no API to determine the path to a secret within a
binding. This required the implementation to be aware of the specified layout,
and environment variable root of the binding specification.  With the addition
of a new SecretFilePath API to libcnb, this knowledge can be removed from the
buildpack.

Signed-off-by: Ben Hale <bhale@vmware.com>