Profiler ProjectID
Previously, the profiler JAR file would extract the project ID from the
Application Credentials.  This was changed at some point and the project ID
now needs to be passed in explicitly.  This change updates the buildpack to
expect $BPL_GOOGLE_STACKDRIVER_PROJECT_ID to be set and pass it to the
profiler agent.

Signed-off-by: Ben Hale <bhale@vmware.com>