Conditional at Launch
Previously, a binding was required at build time in order to contribute the
layers for the Stackdriver components.  We've started to think that requiring
a binding during build is too onerous a restriction and instead environment
variables should be used to indicate that a dependency should be contributed.
The knock-on effect of this change is that we have to move a lot more of the
conditional logic out to launch time including ensuring that if the binding
doesn't exist _nothing_ is contributed.  This change makes those updates to
the buildpack's behavior.

In addition, this change also includes an update that enables the buildpack to
determine if the application is running in GCP by detecting the metadata
server.  In this case, no binding is required at launch time either as the
credentials typically provided by the binding are instead provided by the
metadata server.

Finally, this includes a fix where the provided build plan wasn't anywhere
near exhaustive for combinations that a user need when using the buildpack.

Signed-off-by: Ben Hale <bhale@vmware.com>